rfc6242: add transport Decoder type
Add the RFC6242 Decoder type. This type decodes RFC6242 encoded
transport streams and operates as a filter with support for the
io.Reader and io.WriterTo interfaces.

io.Copy will use this type as an io.WriterTo, permitting
direct (no-copy) transfer and avoiding the use of a pipe.

When using Read, a pipe and a background goroutine consumes
the source reader and makes data available to the Read function.

Benchmark copying from the decoder to /dev/null:
  goos: linux
  goarch: amd64
  BenchmarkDecoderEndOfMessageReader-8    100000  13431 ns/op  99633 B/op  19 allocs/op
  BenchmarkDecoderEndOfMessageWriterTo-8  200000   6561 ns/op  66848 B/op  17 allocs/op
  BenchmarkDecoderChunkedWriterTo-8       200000   6473 ns/op  66720 B/op  17 allocs/op
  BenchmarkDecoderChunkedReader-8         100000  13671 ns/op  99504 B/op  19 allocs/op