rfc6242: add package example for protocol upgrade
Add an example describing framing protocol upgrade when handling a NETCONF
session. Example is written from the client's point of view, with a mock server
input stream given supposed client output.

Shows the interaction between Read() calls made to the Decoder (by calls to the
xml.Decoder's Token() function) and upgrading the framing on the Decoder (and
related session Encoder) object based on the contents of the input.